requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
dnspython==2.2.0
idna==3.3
pip-autoremove==0.10.0
pycodestyle==2.8.0
PyJWT==2.3.0
pymongo==4.0.1
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
